test: test route for response test
