--- test/cases.js
@@ -1,4 +1,0 @@
-module.exports = {
-    text: ['/'],
-    rss: ['/bilibili/ranking/0/3'],
-};
