--- test/index.js
@@ -1,1 +1,0 @@
-require('./response');
