--- test/rules/index.js
@@ -1,7 +1,0 @@
-const status = require('./status');
-const rss = require('./rss');
-
-module.exports = async (response) => {
-    status(response);
-    await rss(response);
-};
