--- test/rules/status.js
@@ -1,3 +1,0 @@
-module.exports = (response) => {
-    expect(response.status).toBe(200);
-};
